Separate definition loading from sync class
Motivation
----------
Allow definition loading to be used by other processes in future work,
such as definition validation.

Modifications
-------------
Moved definition loading logic from sync.js to definition-loader.js.
